requirements.txt
Pillow
argparse